+++ src/main/java/blackjack/Application.java
@@ -1,6 +1,6 @@
 package blackjack;
 
-import blackjack.domain.controller.BlackJackController;
+import blackjack.controller.BlackJackController;
 
 public class Application {
     public static void main(String[] args) {
