+++ src/main/java/blackjack/domain/card/HoldingCard.java
@@ -1,5 +1,7 @@
-package blackjack.domain;
+package blackjack.domain.card;
 
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
 import java.util.ArrayList;
 import java.util.List;
