+++ src/main/java/blackjack/domain/participant/Participant.java
@@ -1,17 +1,20 @@
-package blackjack.domain;
+package blackjack.domain.participant;
+
+import blackjack.domain.card.Card;
+import blackjack.domain.card.HoldingCard;
 
 public abstract class Participant {
-    private String name;
-    private HoldingCard holdingCard;
+    private final String name;
+    private final HoldingCard holdingCard;
 
     public Participant(String name, HoldingCard holdingCard) {
         this.name = name;
         this.holdingCard = holdingCard;
     }
 
-    void receiveCard(Card card){
+    public void receiveCard(Card card) {
         holdingCard.add(card);
-    };
+    }
 
     public String getName() {
         return name;
@@ -33,6 +36,5 @@ public abstract class Participant {
             return 0;
         }
         return score;
-
     }
 }
