--- src/main/java/blackjack/domain/strategy/CardGenerator.java
@@ -1,8 +1,0 @@
-package blackjack.domain.strategy;
-
-import blackjack.domain.Card;
-import java.util.List;
-
-public interface CardGenerator {
-    List<Card> generate();
-}
