+++ src/test/java/blackjack/domain/PlayerTest.java
@@ -1,6 +1,9 @@
 package blackjack.domain;
 
-import blackjack.domain.Player;
+import blackjack.domain.card.Card;
+import blackjack.domain.card.CardNumber;
+import blackjack.domain.card.Symbol;
+import blackjack.domain.participant.Player;
 import java.util.List;
 import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.DisplayName;
